Moved extension to appropriate folder.
