+++ src/DigitalWallet/Common/Extensions/ValidatorExtensions.cs
@@ -1,0 +1,11 @@
+﻿namespace DigitalWallet.Common.Extensions;
+
+internal static class ValidatorExtensions
+{
+    public static RouteHandlerBuilder Validator<T>(this RouteHandlerBuilder handlerBuilder)
+        where T : class
+    {
+        handlerBuilder.AddEndpointFilter<EndpointValidatorFilter<T>>();
+        return handlerBuilder;
+    }
+}
