+++ src/DigitalWallet/Features/MultiCurrency/UpdateRatio/Endpoint.cs
@@ -1,4 +1,5 @@
 ﻿using Carter;
+using DigitalWallet.Common.Extensions;
 
 namespace DigitalWallet.Features.MultiCurrency.UpdateRatio;
