+++ src/DigitalWallet/Features/Transactions/DecreaseWalletBalance/Endpoint.cs
@@ -1,9 +1,10 @@
-﻿ using Carter;
+﻿using Carter;
+using DigitalWallet.Common.Extensions;
 
- namespace DigitalWallet.Features.Transactions.DecreaseWalletBalance;
+namespace DigitalWallet.Features.Transactions.DecreaseWalletBalance;
 
 
- public class Endpoint : ICarterModule
+public class Endpoint : ICarterModule
  {
      public void AddRoutes(IEndpointRouteBuilder app)
      {
